#231 System test now reads result of ls command instead of copied file.
Previously the test copied a file from mesos-local into /tmp on the host. Assertions succeeded but when removing the mesos-local
container we got the following error

[ERROR] 2015-08-24 16:02:20,002 org.apache.mesos.mini.container.AbstractContainer remove - Could not remove container mini_mesos_cluster_-1150231373    com.github.dockerjava.api.InternalServerErrorException: Cannot destroy container 938697f64c3c1502e293a48a13c84e33f2c236a53f31688224ff3fe00c93064b: Driver aufs failed to remove root filesystem 938697f64c3c1502e293a48a13c84e33f2c236a53f31688224ff3fe00c93064b: device or resource busy

It seemed that the Docker client file copy does not clean up properly. When we replaced the file copy with an exec that run ls we can assert the results without error.